requirements.txt
google-generativeai
selenium
